Fix /dev/ptmx rather than excluding makedev
Revert exclusion of makedev from buildd chroots, as it turned out not to
be the problem.  Instead, fix up /dev/ptmx to be a character device node
rather than a symlink to /dev/pts/ptmx, in line with the discussion in
https://bugs.debian.org/817236; I think this is safer than
cherry-picking the fix to debootstrap at this point in a stable release
cycle.

LP: #1844504