fix: Mark grub-pc-bin as manual install to avoid autoremoval (LP: #2045418)
With amd64 build target calling `update-grub` no longer works and fails build with error

```
grub-install: error: /usr/lib/grub/i386-pc/modinfo.sh doesn't exist. Please specify --target or --directory.
```

`/usr/lib/grub/i386-pc/modinfo.sh` is part of the grub-pc-bin package.

We are seeing this now on xenial as grub-pc-bin is now removed during build as part of fixing
bug https://bugs.launchpad.net/ubuntu/+source/livecd-rootfs/+bug/2036195.

To resolve this we mark grub-pc-bin as manually installed so it is not autoremoved.

As we are running grub-install ourselves for a specific target we need to make sure to install that
target first which means this will not be solved by changing dependencies of the grub packages.