update-source-catalog: support version 2 scheme
Version 1 of install-sources.yaml is a top-level list of the sources to
be offered.
Version 2 extends this by placing the list under a top-level key
`sources`, adding a `version` field, and adding a `kernel` field which
supplants the current kernel-meta-package file.  `kernel.default` is
read to know which kernel to use - unless we need to fallback to the
bridge kernel.