fix: Increase CPC disk-image base imagesize to 2.5GB
In the ubuntu-cpc disk-image binary we need to avail of the ever increasing size
of packages. 2.2GB is now just a bit too small leading to `No space
left on device` errors when the binary hits `grub-install`. This commit
increases $imagesize to 2.5GB (in the binary as an override initially
implemented in ecaaf0484).

This commit also runs `df` just after the grub-pc && grub2-common
installs to make for easier debugging in the future.

Refs: LP: #2115811